+++ lib/mongoid/relations/macros.rb
@@ -132,7 +132,7 @@ module Mongoid # :nodoc:
           characterize(name, Referenced::In, options, &block).tap do |meta|
             relate(name, meta)
             reference(meta)
-            autosave(meta)
+            builder(name, meta).creator(name, meta).autosave(meta)
             validates_relation(meta)
           end
         end